refactor: 음식점 정렬/필터링 관련 상수 파일 분리
